+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Option for overwrite the output file if it exists (`-f`)
 - Option for print the version number (`-V`)
 - Human readable total size of processed files in output file information
+### Changed
+- Symbolic links processing removed
+### Fixed
+- 'Too many levels of symbolic links' bug
 
 ## [1.2] - 2019-12-08
 ### Added