+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.3] - Work in progress
+## [1.3] - 2019-12-25
 ### Added
 - Option for disabling any information output (`-q`)
 - Option for overwrite the output file if it exists (`-f`)